requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
environs==7.2.0
marshmallow==3.5.0
psycopg2==2.8.4
python-dotenv==0.11.0
pytz==2019.3
sqlparse==0.3.0
